Update why_use_it.rst
Standardizing on NGINX where appropriate. 

NOTE: The link to the TurboGears mailing list is broken.  It suspect that they are purging material that's over 9 years old. I searched for "nginx" and the results included an item dated 08/28/06 (that's nine years ago today), but not Ippolito's from 08/24/06.

Also, the HowToForge and TimmeHosting links are to the homepage for those sites, not the page with the quote on them.  I don't know how important that is.