Fix pygment error handling
Sphinx re-renders pygments outputs with no syntax highlighting if an
exception is thrown in pygments. Pygments has a filter to throw
exceptions when errors are hit. Ignore these errors because some are due
to Pygments not understanding Lua in Nginx configurations.

Instead errors will have small red boxes around them in the rendered
output, this is controllable with CSS.