Correct recommended fastcgi location regex in Drupal recipe
The fastcgi `location` directive I recommended in #207 matched the new Drupal 8 URL style, `update.php/selection` & c., with a very lax rule, so that future paths of this nature that Drupal 8 might add in the future would also work.  However, this lax regex got in the way of folks with legacy path aliases such as `blog/index.php/legacy-path`, e.g. in instances where they migrated a site from some old CMS that did not use clean URLs, and needed to continue to serve the legacy locations after migration.

This PR therefore recommends a stricter pattern, as suggested by https://insready.com/en/blog/update-your-nginx-config-support-drupal-8, so these sorts of paths will continue to work for those who need them.

Since this situation is somewhat complex, and there isn't really one single solution that is likely to be right for all users for all time (if we include the possibility of future Drupal 8 changes), I have included a fairly long explanatory comment. This could be pruned down, if desired, as the new location should by an acceptable solution for all users for right now.