Systemd: Check configuration on reload
Sending SIGHUP to the NGINX master process is not exactly equivalent to
using `nginx -s reload`. The latter will go through a part of NGINX
initialization procedure and validate the configuration file.

Thus, if an invalid configuration file is provided and NGINX is reloaded
by sending a signal to the master process, systemd is not able to know
that the reload failed (the signal was sent successfully). On the other
hand, using the NGINX CLI option to reload will validate the
configuration file and report the failure before even sending the signal
to the running NGINX process.