Include new learning resource (#443)
We are launching Classpert.com, a meta search engine for online courses. We have a collection of NGINX courses, so I figured it might be useful for newcomers. Please let me know if this kind of proposed edit violates any wiki guidelines (I couldn't find anything about it on the docs)