Update Wordpress recipe
I recently tried to server a wordpress page on Ubuntu 18.04 following the wiki recipe. I only got blank pages and 502 errors, so I searched on the internet and there seemed to be a consensus on these changes I'm proposing to make it work. It did work for me.
Anyway I'm very new to this, so it's necessary someone supervises this.
I hope it's helpful,
J.